# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sports Coach Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a coach
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /coaches/{id}/reviews:
    get:
      operationId: getCoachReviews
      tags: [reviews]
      summary: Get approved reviews for a coach
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /coaches/{id}/photos:
    get:
      operationId: getCoachPhotos
      tags: [photos]
      summary: Get approved photos for a coach
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
    post:
      operationId: uploadCoachPhoto
      tags: [photos]
      summary: Upload a photo for a coach
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhotoBody"
      responses:
        "201":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
...
  /user/preferences:
    get:
      operationId: getUserPreferences
      tags: [user]
      summary: Get personalized category preferences for the current user
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    post:
      operationId: trackCategoryClick
      tags: [user]
      summary: Track a category click for personalization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackClickBody"
      responses:
        "200":
          description: Preference recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
...
components:
  schemas:
...
    Review:
      type: object
      properties:
        id:
          type: integer
        coachId:
          type: integer
        userId:
          type: string
        userName:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - coachId
        - userId
        - rating
        - comment
        - isApproved
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        coachId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        userName:
          type: string
      required:
        - coachId
        - rating
        - comment
    Photo:
      type: object
      properties:
        id:
          type: integer
        coachId:
          type: integer
        imageUrl:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - coachId
        - imageUrl
        - isApproved
        - createdAt
    CreatePhotoBody:
      type: object
      properties:
        imageUrl:
          type: string
      required:
        - imageUrl
...
    UserPreferences:
      type: object
      properties:
        preferredCategories:
          type: array
          items:
            type: string
        clickHistory:
          type: object
          additionalProperties:
            type: integer
      required:
        - preferredCategories
        - clickHistory
    TrackClickBody:
      type: object
      properties:
        category:
          type: string
      required:
        - category
...